chore: Split loaders into separate file
